% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{add_utm}
\alias{add_utm}
\title{Function to append UTM columns to data frame with lat / lon columns}
\usage{
add_utm(
  data,
  ll_cols = c("longitude", "latitude"),
  ll_crs = 4326,
  utm_names = c("X", "Y"),
  utm_crs
)
}
\arguments{
\item{data}{data frame}

\item{ll_cols}{length-2 character vector; names of longitude and latitude columns (in that order)}

\item{ll_crs}{Coordinate reference system of lat-long coordinates}

\item{utm_names}{Names of new UTM columns to be appended to data frame}

\item{utm_crs}{Coordinate reference system for UTM coordinates}
}
\value{
A copy of the input data frame with UTM columns appended
}
\description{
The bulk of this function is duplicated from sdmTMB::add_utm_columns
}
