% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{AUC}
\alias{AUC}
\title{Wrapper function to extract AUC using dismo::evaluate}
\usage{
AUC(obs, fit)
}
\arguments{
\item{obs}{Observed presence-absence (0-1) data}

\item{fit}{Fitted probabilities}
}
\value{
AUC value
}
\description{
Wrapper function to extract AUC using dismo::evaluate
}
