% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roms_level2.R
\name{bias_correct}
\alias{bias_correct}
\title{Delta-correct ROMS level 2 outputs}
\usage{
bias_correct(x, hindcast, historical, lower = NA, upper = NA)
}
\arguments{
\item{x}{ROMS level 2 \code{stars} object returned by \code{get_level2}}

\item{hindcast}{Weekly means for ROMS level 2 hindcast, returned by \code{weight_weekly}}

\item{historical}{Weekly means for ROMS level 2 historical run, returned by \code{weight_weekly}}

\item{lower}{If applicable, lower threshold for returned variable}

\item{upper}{If applicable, upper thresdhold for returned variable}
}
\value{
A \code{stars} object
}
\description{
Delta-correct ROMS level 2 outputs
}
