% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{cloglog}
\alias{cloglog}
\title{complimentary log-log link / transform}
\usage{
cloglog(x, inverse = FALSE)
}
\arguments{
\item{x}{Numeric}

\item{inverse}{if TRUE, use inverse link}
}
\value{
A numeric vector
}
\description{
complimentary log-log link / transform
}
