% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roms_level2.R
\name{get_level2}
\alias{get_level2}
\title{Download Bering 10K level 2 gridded ROMS output}
\usage{
get_level2(
  var,
  type = c("hindcast", "forecast", "historical"),
  scenario = c("SSP126", "SSP585", "RCP45", "RCP85"),
  start = NA,
  end = NA,
  version = c("CMIP6", "CMIP5"),
  earth_model = c("GFDL", "CESM", "MIROC"),
  crop_ebs = TRUE,
  write_dir = "."
)
}
\arguments{
\item{var}{variable name}

\item{type}{"hindcast", "forecast", or "historical" run}

\item{scenario}{"SSP126" or "SSP585" for CMIP6 outputs, "RCP45" or "RCP85" for CMIP5 outputs}

\item{start}{integer start year. NA defaults to all available years.}

\item{end}{integer end year. NA defaults to all available years.}

\item{version}{"CMIP6" or "CMIP5"}

\item{earth_model}{"GFDL", "CESM", or "MIROC"}

\item{crop_ebs}{if TRUE, crop to EBS/NBS survey boundary}

\item{write_dir}{path to write downloaded raster. if NA, object is not written to disk.}
}
\value{
A \code{stars} object
}
\description{
Download Bering 10K level 2 gridded ROMS output
}
