% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_sediment}
\alias{get_sediment}
\title{NOAA 15-arcsecond EBS digitial bedrock elevation model}
\usage{
get_sediment()
}
\value{
A "stars" object
}
\description{
NOAA 15-arcsecond EBS digitial bedrock elevation model
}
