% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{loc_collocfn}
\alias{loc_collocfn}
\alias{biomass_overlapfn}
\alias{hurlbert_overlapfn}
\alias{AB_overlapfn}
\alias{bhatta_coeffn}
\alias{glob_collocfn}
\alias{area_overlapfn}
\alias{range_overlapfn}
\title{Predator-prey overlap}
\usage{
loc_collocfn(prey, pred)

biomass_overlapfn(prey, pred)

hurlbert_overlapfn(prey, pred, area)

AB_overlapfn(prey, pred)

bhatta_coeffn(prey, pred)

glob_collocfn(prey_x, prey_y, prey, pred_x, pred_y, pred)

area_overlapfn(prey, pred, area)

range_overlapfn(prey, pred, area)
}
\arguments{
\item{prey}{Vector of prey density / biomass values}

\item{pred}{Vector of predator density / biomass values}

\item{area}{area of survey corresponding to density / biomass estimates (no units)}
}
\value{
Numeric overlap value
}
\description{
Functions to compute spatial predator-prey overlap
}
\details{
Selected functions to compute spatial predator-prey overlap from estimated predator and
prey densities. Code is adapted from Carroll et al. (2019), 'A review of methods
for quantifying predator-prey overlap,' Global Ecology and Biogeography. Please see
Carroll et al. for detailed descriptions of provided metrics and their ecological interpretations:
\itemize{
\item \code{loc_collocfn}: The local index of collocation
\item \code{glob_collocfn}: The global index of collocation
\item \code{biomass_overlapfn}: Biomass-weighted overlap
\item \code{hurlbert_overlapfn}: Hurlbert's overlap
\item \code{AB_overlapfn}: AB ratio
\item \code{area_overlapfn}: Area overlap
\item \code{range_overlapfn}: Range overlap
}
}
