% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smoother.R
\name{plot_smoother}
\alias{plot_smoother}
\title{Plot smooth terms for mgcv::gam objects}
\usage{
plot_smoother(model, s, scale = c("link", "response"), rug = TRUE, ...)
}
\arguments{
\item{model}{fitted model object}

\item{s}{name of predictor to plot smooth term for}

\item{scale}{if "response", inverse link function will be applied}

\item{rug}{show data rug at bottom of plot?}

\item{...}{additional arguments to geom_ribbon. Providing args (e.g. fill) will reset defaults for other args (alpha, color, linetype).}
}
\value{
a ggplot object
}
\description{
Plot smooth terms for mgcv::gam objects
}
