% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smoother.R
\name{plot_smoother_2D}
\alias{plot_smoother_2D}
\title{Plot 2D smooth terms for mgcv::gam objects}
\usage{
plot_smoother_2D(model, s, se = TRUE, rug = TRUE, crop_ebs = TRUE)
}
\arguments{
\item{model}{fitted model object}

\item{s}{label corresponding to smooth term (e.g. "s(x, y, k = ..., ...)") to plot}

\item{se}{Logical - should a seprate plot for smoother standard error be returned?}

\item{rug}{Logical - should points corresponding to observed data be plotted?}

\item{crop_ebs}{Logical - should smoother be restricted to EBS survey region?}
}
\value{
A ggplot2 object
}
\description{
Plot 2D smooth terms for mgcv::gam objects
}
