% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{rescale_01}
\alias{rescale_01}
\title{Simple function to linearly rescale variables to 0-1 interval}
\usage{
rescale_01(x, na.rm = TRUE)
}
\arguments{
\item{x}{Variable to rescale}

\item{na.rm}{Defaults to TRUE}
}
\value{
A numeric vector
}
\description{
Simple function to linearly rescale variables to 0-1 interval
}
