% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rotate_lon}
\alias{rotate_lon}
\title{Simple function to convert longitude from 0/360 to -180/180 and vice-versa}
\usage{
rotate_lon(x, from = c("-180/180", "0/360"))
}
\arguments{
\item{x}{A numeric vector of longitudes}

\item{from}{Whether to convert from -180/180 or 0/360}
}
\value{
A numeric vector
}
\description{
Simple function to convert longitude from 0/360 to -180/180 and vice-versa
}
