% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{solve_weights}
\alias{solve_weights}
\title{Obtain stacked weights}
\usage{
solve_weights(yi, yhat)
}
\arguments{
\item{yi}{Length-n vector of observations}

\item{yhat}{n by m matrix of n predictions for m models}
}
\value{
A vector of length m corresponding to model weights
}
\description{
Obtain stacked weights
}
