% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roms_level2.R
\name{weight_weeks}
\alias{weight_weeks}
\title{Compute week-of-year means for ROMS output}
\usage{
weight_weeks(x, start = NA, end = NA)
}
\arguments{
\item{x}{A multiband \code{stars} object returned by \code{get_level2}}

\item{start}{integer start year. NA defaults to all available years.}

\item{end}{integer end year. NA defaults to all available years.}
}
\value{
A \code{stars} object
}
\description{
Compute week-of-year means for ROMS output
}
