% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{weighted_se}
\alias{weighted_se}
\title{Compute variance of ensemble mean}
\usage{
weighted_se(X, S, w)
}
\arguments{
\item{X}{n x m matrix of predictions for n data points across m models}

\item{S}{n x m matrix of estimate standard errors}

\item{w}{m-length vector of model weights}
}
\value{
A length n vector of ensemble mean standard errors
}
\description{
Compute variance of ensemble mean
}
